+++ src/app/components/index.ts
@@ -15,3 +15,4 @@ import "./new-version-hint.component";
 import "./signal-canvas-markup-canvas.component";
 import "./level-canvas-markup-canvas.component";
 import "./epg-grid-canvas.component";
+import "./dots-canvas.component";