+++ src/NEWS.md
@@ -9,7 +9,7 @@
 - font scaling factor now also applied to hierarchy path canvas
 - moved emuwebappDesign from sass to scss and made it importable with .d.ts file
 - removed default_emuwebappDesign.json file to have single source of truth in scss file
-- epg canvas now a component
+- epg and dot canvas now components
 
 ### bug fixes